requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
XlsxWriter==3.2.0
